+++ src/the_music_tree_genre_kit/criteria/children/tag/TagManager.py
@@ -1,8 +1,0 @@
-from the_music_tree_genre_kit.criteria.AbstractCriteriaManager import AbstractCriteriaManager
-from the_music_tree_genre_kit.criteria.type.CriteriaType import CriteriaType
-from the_music_tree_genre_kit.criteria.type.CriteriaTypePks import CriteriaTypePks
-
-
-class TagManager(AbstractCriteriaManager):
-    def _get_criteria_type(self) -> CriteriaType:
-        return CriteriaType(pk=CriteriaTypePks.TAG)